Basic inventory (#6)
API to read and store the inventory from a remote source. This is
currently limited to PMM.

* Added sep.inventory.api along with data models. The inventory is
initialised from the remote source when the local data is emtpy.
Additionally, the inventory can be refreshed by request. There is
partial support to create and update items in the inventory directly via
the API, only affecting local storage, which may be overwritten if
changing an item that is present in the remote source